@@ -5,9 +5,8 @@ import {
   HttpStatus,
 } from '@nestjs/common';
 import { HttpException } from '@nestjs/common';
-import { Observable } from 'rxjs/Observable';
+import { Observable, throwError } from 'rxjs';
 import { catchError } from 'rxjs/operators';
-import { _throw } from 'rxjs/observable/throw';
 
 @Injectable()
 export class ErrorsInterceptor implements NestInterceptor {
@@ -17,7 +16,7 @@ export class ErrorsInterceptor implements NestInterceptor {
   ): Observable<any> {
     return call$.pipe(
       catchError(err =>
-        _throw(new HttpException('Message', HttpStatus.BAD_GATEWAY)),
+        throwError(new HttpException('Message', HttpStatus.BAD_GATEWAY)),
       ),
     );
   }
