@@ -1,11 +1,10 @@
 import { Catch, RpcExceptionFilter } from '@nestjs/common';
-import { Observable } from 'rxjs/Observable';
+import { Observable, throwError } from 'rxjs';
 import { RpcException } from '@nestjs/microservices';
-import { _throw } from 'rxjs/observable/throw';
 
 @Catch(RpcException)
 export class ExceptionFilter implements RpcExceptionFilter {
   catch(exception: RpcException): Observable<any> {
-    return _throw(exception.getError());
+    return throwError(exception.getError());
   }
 }
